+++ src/main/java/Controller/MainGameEngine.java
@@ -3,6 +3,7 @@ package Controller;
 import Models.CurrentState;
 import Models.Orders;
 import Models.Player;
+import Utils.CommandHandler;
 import Views.MapView;
 
 import java.io.BufferedReader;
