+++ src/main/java/Utils/CommandHandler.java
@@ -1,4 +1,4 @@
-package Controller;
+package Utils;
 
 import java.util.*;
